requirements.txt
flask
gunicorn
quant-platform-kit @ git+https://github.com/QuantStrategyLab/QuantPlatformKit.git@v0.7.19
us-equity-strategies @ git+https://github.com/QuantStrategyLab/UsEquityStrategies.git@v0.7.31
pandas
requests
pytz
pandas_market_calendars
yfinance
google-auth
google-cloud-secret-manager
google-cloud-storage
google-cloud-firestore
